--- src/PUGX/Badge/Image/ImageInterface.php
@@ -1,21 +1,0 @@
-<?php
-
-/*
- * This file is part of the badge-poser package.
- *
- * (c) PUGX <http://pugx.github.io/>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace PUGX\Badge\Image;
-
-/**
- * @author Claudio D'Alicandro <claudio.dalicandro@gmail.com>
- * @author Giulio De Donato <liuggio@gmail.com>
- */
-interface ImageInterface
-{
-    public function __toString();
-}
