+++ src/PUGX/Badge/Image/Tests/Factory/SvgImageFactoryTest.php
@@ -11,16 +11,16 @@
 
 namespace PUGX\Badge\Image\Tests\Factory;
 
-use PUGX\Badge\Image\Factory\SvgImageFactory;
+use PUGX\Badge\Image\Factory\PoserImageFactory;
 
 /**
- * Class SvgImageFactoryTest
+ * Class PoserImageFactoryTest
  *
  * @author Claudio D'Alicandro <claudio.dalicandro@gmail.com>
  */
-class SvgImageFactoryTest extends \PHPUnit_Framework_TestCase
+class PoserImageFactoryTest extends \PHPUnit_Framework_TestCase
 {
-    /** @var SvgImageFactory $imageFactory */
+    /** @var PoserImageFactory $imageFactory */
     private $imageFactory;
 
     public function setUp()
@@ -32,13 +32,13 @@ class SvgImageFactoryTest extends \PHPUnit_Framework_TestCase
             ->method('generate')
             ->will($this->returnValue('<svg \>'));
 
-        $this->imageFactory = new SvgImageFactory($shieldGenerator);
+        $this->imageFactory = new PoserImageFactory($shieldGenerator);
     }
 
     public function testShouldCreateDownloadsImage()
     {
         $image = $this->imageFactory->createDownloadsImage('test', 'test', 'red');
 
-        $this->assertInstanceOf('PUGX\Badge\Image\ImageInterface', $image);
+        $this->assertInstanceOf('PUGX\Badge\Image\Image', $image);
     }
 }
