--- src/PUGX/Badge/Package/ApiPackageInterface.php
@@ -1,39 +1,0 @@
-<?php
-
-/*
- * This file is part of the badge-poser package.
- *
- * (c) PUGX <http://pugx.github.io/>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace PUGX\Badge\Package;
-
-/**
- * Class ApiPackageInterface
- *
- * @author Giulio De Donato <liuggio@gmail.com>
- */
-interface ApiPackageInterface
-{
-    public function getName();
-
-    public function getDescription();
-
-    public function getDownloads();
-
-    public function getFavers();
-
-    public function getMaintainers();
-
-    public function getTime();
-
-    public function getRepository();
-
-    public function getType();
-
-    public function getVersions();
-
-}
