+++ src/PUGX/Badge/Package/Package.php
@@ -14,12 +14,13 @@ namespace PUGX\Badge\Package;
 use Packagist\Api\Result\Package as ApiPackage;
 
 /**
- * Class Package
+ * Class Package, decorates the Packagist Package.
  *
  * @author Giulio De Donato <liuggio@gmail.com>
  */
-class Package implements PackageInterface, ApiPackageInterface
+class Package
 {
+    private static $modifierRegex = '[._-]?(?:(stable|beta|b|RC|alpha|a|patch|pl|p)(?:[.-]?(\d+))?)?([.-]?dev)?';
     private $license;
     private $originalObject;
     private $latestStableVersion = null;
@@ -27,19 +28,128 @@ class Package implements PackageInterface, ApiPackageInterface
     private $latestStableVersionNormalized = null;
     private $latestUnstableVersionNormalized = null;
 
-    public function setLicense($license)
+    private function __construct(ApiPackage $apiPackage)
     {
-        $this->license = $license;
+       $this->setOriginalObject($apiPackage);
+       $this->calculateLatestVersions();
+    }
+
+    /**
+     * Create a new Package decorated with the Api Package.
+     *
+     * @param ApiPackage $apiPackage
+     *
+     * @return PackageInterface
+     */
+    public static function createFromApi(ApiPackage $apiPackage)
+    {
+        $package = new self($apiPackage);
+
+        return $package;
+    }
+
+    /**
+     * Set the latest Stable and the latest Unstable version from a Package.
+     *
+     * @return Package
+     */
+    private function calculateLatestVersions()
+    {
+        $versions = $this->getVersions();
+
+        foreach ($versions as $name => $version) {
+
+            $currentVersionName = $version->getVersion();
+            $versionNormalized = $version->getVersionNormalized();
+
+            $aliases = $this->getBranchAliases($version);
+            if (null !== $aliases && array_key_exists($currentVersionName, $aliases)) {
+                $currentVersionName = $aliases[$currentVersionName];
+            }
+
+            $functionName = 'Unstable';
+            if ('stable' == self::parseStability($currentVersionName)) {
+                $functionName = 'Stable';
+            }
+
+            if (version_compare($versionNormalized, $this->{'getLatest' . $functionName . 'VersionNormalized'}()) > 0) {
+                $this->{'setLatest' . $functionName . 'Version'}($currentVersionName);
+                $this->{'setLatest' . $functionName . 'VersionNormalized'}($versionNormalized);
+
+                $license = $version->getLicense();
+                if (is_array($license) && count($license)>0) {
+                    $license = implode(',',$license);
+                }
+                $this->setLicense($license);
+            }
+        }
+
+        return $this;
+    }
+
+    /**
+     * Get all the branch aliases.
+     *
+     * @param ApiPackage\Version $version
+     *
+     * @return null|array
+     */
+    private function getBranchAliases(ApiPackage\Version $version)
+    {
+        $extra = $version->getExtra();
+        if (null !== $extra
+            && null !== $extra["branch-alias"]
+            && is_array($extra["branch-alias"])
+        ) {
+            return $extra["branch-alias"];
+        }
+
+        return null;
+    }
+
+    /**
+     * Returns the stability of a version.
+     *
+     * This function is part of Composer.
+     *
+     * (c) Nils Adermann <naderman@naderman.de>
+     * Jordi Boggiano <j.boggiano@seld.be>
+     *
+     * @param string $version
+     *
+     * @return string
+     */
+    public static function parseStability($version)
+    {
+        $version = preg_replace('{#.+$}i', '', $version);
+
+        if ('dev-' === substr($version, 0, 4) || '-dev' === substr($version, -4)) {
+            return 'dev';
+        }
+
+        preg_match('{' . self::$modifierRegex . '$}i', strtolower($version), $match);
+        if (!empty($match[3])) {
+            return 'dev';
+        }
+
+        if (!empty($match[1])) {
+            if ('beta' === $match[1] || 'b' === $match[1]) {
+                return 'beta';
+            }
+            if ('alpha' === $match[1] || 'a' === $match[1]) {
+                return 'alpha';
+            }
+            if ('rc' === $match[1]) {
+                return 'RC';
+            }
+        }
+
+        return 'stable';
     }
 
     public function getLicense()
     {
         return $this->license;
-    }
-
-    public function setLatestStableVersion($version)
-    {
-        $this->latestStableVersion = $version;
     }
 
     public function getLatestStableVersion()
@@ -52,22 +162,9 @@ class Package implements PackageInterface, ApiPackageInterface
         return isset($this->latestStableVersion);
     }
 
-    public function setLatestUnstableVersion($version)
-    {
-        $this->latestUnstableVersion = $version;
-    }
-
     public function getLatestUnstableVersion()
     {
         return $this->latestUnstableVersion;
-    }
-
-    /**
-     * @param null $latestStableVersionNormalized
-     */
-    public function setLatestStableVersionNormalized($latestStableVersionNormalized)
-    {
-        $this->latestStableVersionNormalized = $latestStableVersionNormalized;
     }
 
     /**
@@ -78,17 +175,6 @@ class Package implements PackageInterface, ApiPackageInterface
         return $this->latestStableVersionNormalized;
     }
 
-    /**
-     * @param null $latestUnstableVersionNormalized
-     */
-    public function setLatestUnstableVersionNormalized($latestUnstableVersionNormalized)
-    {
-        $this->latestUnstableVersionNormalized = $latestUnstableVersionNormalized;
-    }
-
-    /**
-     * @return null
-     */
     public function getLatestUnstableVersionNormalized()
     {
         return $this->latestUnstableVersionNormalized;
@@ -99,14 +185,34 @@ class Package implements PackageInterface, ApiPackageInterface
         return isset($this->latestUnstableVersion);
     }
 
-    public function setOriginalObject(ApiPackage $originalObject)
-    {
-        $this->originalObject = $originalObject;
-    }
-
     public function getOriginalObject()
     {
         return $this->originalObject;
+    }
+
+    private function setLatestUnstableVersionNormalized($latestUnstableVersionNormalized)
+    {
+        $this->latestUnstableVersionNormalized = $latestUnstableVersionNormalized;
+    }
+
+    private function setLicense($license)
+    {
+        $this->license = $license;
+    }
+
+    private function setLatestStableVersion($version)
+    {
+        $this->latestStableVersion = $version;
+    }
+
+    private function setLatestUnstableVersion($version)
+    {
+        $this->latestUnstableVersion = $version;
+    }
+
+    private function setLatestStableVersionNormalized($latestStableVersionNormalized)
+    {
+        $this->latestStableVersionNormalized = $latestStableVersionNormalized;
     }
 
     // original object's property
@@ -154,5 +260,10 @@ class Package implements PackageInterface, ApiPackageInterface
     public function getVersions()
     {
         return $this->getOriginalObject()->getVersions();
+    }
+
+    private function setOriginalObject(ApiPackage $originalObject)
+    {
+        $this->originalObject = $originalObject;
     }
 }
