+++ src/PUGX/BadgeBundle/Resources/config/services.xml
@@ -38,6 +38,5 @@
 
         <service id="normalizer" class="%normalizer.class%" public="false">
         </service>
-
     </services>
 </container>
