+++ src/PUGX/BadgeBundle/Resources/views/Page/home.html.twig
@@ -26,7 +26,7 @@
 
     <div class="row-fluid bundles">
         <div class="span12"><p class="text-center">
-                More than {{ total_access }} badges served<hr class="styled"></div>
+                {{ badges_served_svg|raw }}<hr class="styled"></div>
     </div>
     <div class="jumbotron">
         <h1>Badge Poser</h1>
