--- src/PUGX/BadgeBundle/Resources/views/Svg/defaultShield.svg.twig
@@ -1,18 +1,0 @@
-<svg xmlns="http://www.w3.org/2000/svg" width="{{ totalWidth }}" height="18">
-    <linearGradient id="smooth" x2="0" y2="100%">
-        <stop offset="0"  stop-color="#fff" stop-opacity=".7"/>
-        <stop offset=".1" stop-color="#aaa" stop-opacity=".1"/>
-        <stop offset=".9" stop-color="#000" stop-opacity=".3"/>
-        <stop offset="1"  stop-color="#000" stop-opacity=".5"/>
-    </linearGradient>
-    <rect rx="4" width="{{ totalWidth }}" height="18" fill="{{ vendorColor }}"/>
-    <rect rx="4" x="{{ vendorWidth }}" width="{{ valueWidth }}" height="18" fill="{{ valueColor }}"/>
-    <rect x="{{ vendorWidth }}" width="4" height="18" fill="{{ valueColor }}"/>
-    <rect rx="4" width="{{ totalWidth }}" height="18" fill="url(#smooth)"/>
-    <g fill="#fff" text-anchor="middle" font-family="DejaVu Sans,Verdana,Geneva,sans-serif" font-size="11">
-        <text x="{{ vendorStartPosition }}" y="13" fill="#010101" fill-opacity=".3">{{ vendor }}</text>
-        <text x="{{ vendorStartPosition }}" y="12">{{ vendor }}</text>
-        <text x="{{ valueStartPosition }}" y="13" fill="#010101" fill-opacity=".3">{{ value }}</text>
-        <text x="{{ valueStartPosition }}" y="12">{{ value }}</text>
-    </g>
-</svg>
